My Sites: move add-new-site button to bottom of sidebar for single site users.
