+++ assets/stylesheets/_components.scss
@@ -237,6 +237,7 @@
 @import 'my-sites/sharing/connections/account-dialog';
 @import 'my-sites/sharing/connections/account-dialog-account';
 @import 'my-sites/sharing/connections/services-group';
+@import 'my-sites/sidebar/style';
 @import 'my-sites/sidebar-navigation/style';
 @import 'my-sites/site-indicator/style';
 @import 'my-sites/site-settings/style';
