+++ client/my-sites/current-site/style.scss
@@ -42,26 +42,6 @@
 	}
 }
 
-.current-site__add-new-site {
-	background: $gray-light;
-	border-bottom: 1px solid lighten( $gray, 25% );
-	display: block;
-	padding: 4px 8px;
-	margin: 0;
-	box-sizing: border-box;
-	position: relative;
-
-	@include breakpoint( "<660px" ) {
-		background-color: $gray-light;
-		padding: 18px 16px 18px 53px;
-	}
-
-	&:hover {
-		background-color: $gray-light;
-		color: $blue-medium;
-	}
-}
-
 .current-site__switch-sites {
 	background: lighten( $gray, 30% );
 	border-bottom: 1px solid lighten( $gray, 25% );
