+++ client/my-sites/sidebar/style.scss
@@ -1,0 +1,16 @@
+.my-sites-sidebar__add-new-site.button.is-borderless.is-compact {
+	background: lighten( $gray, 30% );
+	border-top: 1px solid lighten( $gray, 25% );
+	display: block;
+	padding: 14px 8px;
+	margin: 0;
+	box-sizing: border-box;
+	position: absolute;
+		bottom: 0;
+	width: 100%;
+
+	@include breakpoint( "<660px" ) {
+		padding: 18px 16px 18px 53px;
+		position: static;
+	}
+}
